+++ src/main/resources/db/changelog/009__add_role_to_workspace_invitations.sql
@@ -1,0 +1,5 @@
+--liquibase formatted sql
+--changeset matigfv:9
+
+ALTER TABLE workspace_invitations
+    ADD COLUMN role VARCHAR(50) NOT NULL DEFAULT 'COLLABORATOR';